Add new events helper file for all shared event code
